use same mem

+++ src/ref_cpu.rs
@@ -23,7 +23,7 @@ pub trait Mem {
   fn mem_read_u16(&self, pos: u16) -> u16 {
     let lo = self.mem_read(pos) as u16;
     let hi = self.mem_read(pos + 1) as u16;
-    (hi << 8) | (lo as u16)
+    hi << 8 | lo
   }
 
   fn mem_write_u16(&mut self, pos: u16, data: u16) {
